Add support for specifying registration ID
This adds support for specifying the registration ID on the command line. It also extracts out the configuration handling into its own file.
